requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
clerk-backend-api
openai
svix